requirements.txt
# XXXX Agent Dependencies

claude-agent-sdk>=0.1.26
bedrock-agentcore>=0.1.0
python-dotenv>=1.0.0
boto3>=1.34.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
httpx>=0.26.0
mcp>=1.0.0
redis[hiredis]>=5.0.0

# Observability: Manual OTEL SDK for AgentCore Evaluations
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
aws-opentelemetry-distro>=0.10.0
aws-requests-auth
requests
